+++ app/services/by_month_project_fetcher_service.rb
@@ -4,7 +4,7 @@ class ByMonthProjectFetcherService
   end
 
   def confirmed(month, year)
-    projects = Project.includes(:tasks_data).confirmed.filtered_by_significant_date(month, year)
+    projects = Project.confirmed.filtered_by_significant_date(month, year)
 
     AcademiesApiPreFetcherService.new.call!(projects) if @pre_fetch_academies_api
     sort_by_conditions_met_and_name(projects)
@@ -19,9 +19,9 @@ class ByMonthProjectFetcherService
 
   def confirmed_openers_by_team(month, year, team)
     projects = if team.eql?("regional_casework_services")
-      Project.assigned_to_regional_caseworker_team.includes(:tasks_data).confirmed.filtered_by_significant_date(month, year)
+      Project.assigned_to_regional_caseworker_team.confirmed.filtered_by_significant_date(month, year)
     else
-      Project.not_assigned_to_regional_caseworker_team.by_region(team).includes(:tasks_data).confirmed.filtered_by_significant_date(month, year)
+      Project.not_assigned_to_regional_caseworker_team.by_region(team).confirmed.filtered_by_significant_date(month, year)
     end
 
     AcademiesApiPreFetcherService.new.call!(projects) if @pre_fetch_academies_api
@@ -30,9 +30,9 @@ class ByMonthProjectFetcherService
 
   def revised_openers_by_team(month, year, team)
     projects = if team.eql?("regional_casework_services")
-      Project.assigned_to_regional_caseworker_team.includes(:tasks_data).significant_date_revised_from(month, year)
+      Project.assigned_to_regional_caseworker_team.significant_date_revised_from(month, year)
     else
-      Project.not_assigned_to_regional_caseworker_team.by_region(team).includes(:tasks_data).significant_date_revised_from(month, year)
+      Project.not_assigned_to_regional_caseworker_team.by_region(team).significant_date_revised_from(month, year)
     end
 
     AcademiesApiPreFetcherService.new.call!(projects) if @pre_fetch_academies_api