+++ spec/features/conversions/users_can_complete_conversion_tasks_spec.rb
@@ -13,7 +13,6 @@ RSpec.feature "Users can complete conversion tasks" do
   mandatory_tasks = %w[
     check_accuracy_of_higher_needs
     commercial_transfer_agreement
-    conditions_met
     land_questionnaire land_registry
     receive_grant_payment_certificate redact_and_send
     school_completed share_information single_worksheet
@@ -31,6 +30,7 @@ RSpec.feature "Users can complete conversion tasks" do
     funding_agreement_contact
     risk_protection_arrangement
     sponsored_support_grant
+    conditions_met
   ]
 
   it "confirms we are checking all tasks" do
@@ -204,6 +204,19 @@ RSpec.feature "Users can complete conversion tasks" do
       click_on I18n.t("task_list.continue_button.text")
 
       expect(project.reload.tasks_data.sponsored_support_grant_type).to eq "full_sponsored"
+    end
+  end
+
+  describe "the confirm all conditions met task" do
+    let(:project) { create(:conversion_project, assigned_to: user) }
+
+    scenario "they can confirm all conditions have been met" do
+      visit project_tasks_path(project)
+      click_on "Confirm all conditions have been met"
+      page.find_all(".govuk-checkboxes__input").each { |checkbox| checkbox.click }
+      click_on I18n.t("task_list.continue_button.text")
+
+      expect(project.reload.all_conditions_met).to be true
     end
   end